Refactor: payment 도메인 제거 후, bankingService, costCalculator 로 세분화
