+++ module-domain/src/main/java/com/programmers/dev/exception/BankingException.java
@@ -1,0 +1,5 @@
+package com.programmers.dev.exception;
+
+public class BankingException extends RuntimeException{
+
+}
