--- module-domain/src/main/java/com/programmers/dev/payment/application/PaymentService.java
@@ -1,14 +1,0 @@
-package com.programmers.dev.payment.application;
-
-
-import com.programmers.dev.user.domain.User;
-import org.springframework.stereotype.Service;
-
-
-@Service
-public class PaymentService {
-
-    public void pay(User user, Long money) {
-        user.withdraw(money);
-    }
-}
